+++ js/settings.js
@@ -240,7 +240,7 @@
             });
         });
 
-        $(document).on("click", ".onlyoffice-template-preview", function (event) {
+        $(document).on("click", ".onlyoffice-template-item p", function (event) {
             var item = $(event.target).parents(".onlyoffice-template-item");
             var templateId = $(item).attr("data-id");
 
@@ -251,6 +251,18 @@
             });
 
             window.open(url);
+        });
+
+        $(document).on("click", ".onlyoffice-template-download", function (event) {
+            var item = $(event.target).parents(".onlyoffice-template-item");
+            var templateId = $(item).attr("data-id");
+
+            var downloadLink = OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/downloadas?fileId={fileId}&template={template}",{
+                fileId: templateId,
+                template: "true"
+            });
+
+            location.href = downloadLink;
         });
     });