+++ appinfo/application.php
@@ -33,6 +33,7 @@ use OCA\Onlyoffice\Controller\TemplateController;
 use OCA\Onlyoffice\Crypt;
 use OCA\Onlyoffice\Hookhandler;
 use OCA\Onlyoffice\Hooks;
+use OCA\Onlyoffice\Notifier;
 use OCA\Onlyoffice\Preview;
 
 class Application extends App {
@@ -101,6 +102,22 @@ class Application extends App {
                 return $container->query(Preview::class);
             });
         }
+
+        $notificationManager = \OC::$server->getNotificationManager();
+        $notificationManager->registerNotifier(function () use ($appName) {
+            return new Notifier(
+                $appName,
+                \OC::$server->getL10NFactory(),
+                \OC::$server->getURLGenerator(),
+                \OC::$server->getLogger(),
+                \OC::$server->getUserManager()
+            );
+        }, function () use ($appName) {
+            return [
+                "id" => $appName,
+                "name" => $appName,
+            ];
+        });
 
         $container->registerService("L10N", function ($c) {
             return $c->query("ServerContainer")->getL10N($c->query("AppName"));