+++ js/editor.js
@@ -140,6 +140,8 @@
                         config.events.onRequestInsertImage = OCA.Onlyoffice.onRequestInsertImage;
                         config.events.onRequestMailMergeRecipients = OCA.Onlyoffice.onRequestMailMergeRecipients;
                         config.events.onRequestCompareFile = OCA.Onlyoffice.onRequestCompareFile;
+                        config.events.onRequestUsers = OCA.Onlyoffice.onRequestUsers;
+                        config.events.onRequestSendNotify = OCA.Onlyoffice.onRequestSendNotify;
                         config.events.onMetaChange = OCA.Onlyoffice.onMetaChange;
 
                         if (!OCA.Onlyoffice.filePath) {
@@ -451,6 +453,44 @@
         }
 
         OCA.Onlyoffice.docEditor.setActionLink(url);
+    };
+
+    OCA.Onlyoffice.onRequestUsers = function (event) {
+        $.get(OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/ajax/users"),
+        function onSuccess(response) {
+            OCA.Onlyoffice.docEditor.setUsers({
+                "users": response
+            });
+        });
+    };
+
+    OCA.Onlyoffice.onRequestSendNotify = function (event) {
+        var actionLink = event.data.actionLink;
+        var comment = event.data.message;
+        var emails = event.data.emails;
+
+        var fileId = OCA.Onlyoffice.fileId;
+
+        $.post(OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/ajax/mention"),
+            {
+                fileId: fileId,
+                anchor: JSON.stringify(actionLink),
+                comment: comment,
+                emails: emails
+            },
+            function onSuccess(response) {
+                if (response.error) {
+                    OC.Notification.show(response.error, {
+                        type: "error",
+                        timeout: 3
+                    });
+                    return;
+                }
+
+                OC.Notification.show(response.message, {
+                    timeout: 3
+                });
+            });
     };
 
     OCA.Onlyoffice.onMetaChange = function (event) {